[Refactor] modify search 컴포넌트 분할
