+++ client/src/components/Menu/Search.tsx
@@ -1,7 +1,7 @@
 import { Card } from '@material-tailwind/react';
-import { Input, Select, Option } from '@material-tailwind/react';
-import { ChangeEvent, FormEvent, useEffect, useState } from 'react';
-import { AiOutlineSearch } from 'react-icons/ai';
+import { useEffect, useState } from 'react';
+import SearchForm from './SearchForm';
+import SearchSelect from './SearchSelect';
 
 type props = {
   handlerSearch: () => void;
@@ -15,25 +15,6 @@ export const Search = ({ handlerSearch }: props) => {
     setInputText(searchList.join(','));
   }, [searchList]);
 
-  const handlerSearching = (e: React.MouseEvent<HTMLLIElement>) => {
-    const target = e.target as HTMLLIElement;
-    setSearchList((pre) => {
-      const updatedSearch = [...pre, target.innerHTML];
-      const newData: string[] = Array.from(new Set(updatedSearch));
-      return newData;
-    });
-  };
-
-  const handleFormSubmit = (e: FormEvent<HTMLButtonElement>) => {
-    e.preventDefault();
-    setSearchList(inputText.split(','));
-  };
-
-  const handlerInput = (e: ChangeEvent<HTMLInputElement>) => {
-    const newText = e.target.value;
-    setInputText(newText);
-  };
-
   return (
     <>
       <div
@@ -41,36 +22,12 @@ export const Search = ({ handlerSearch }: props) => {
         onClick={handlerSearch}
       ></div>
       <Card className="w-[350px] p-2 shadow-xl top-3 absolute m-3">
-        <form id="myForm">
-          <div className="mb-3">
-            <Select label="과목" size="lg">
-              <Option onClick={handlerSearching}>국어</Option>
-              <Option onClick={handlerSearching}>수학</Option>
-            </Select>
-          </div>
-          <div className="mb-3">
-            <Select label="지역" size="lg">
-              <Option onClick={handlerSearching}>서울</Option>
-              <Option onClick={handlerSearching}>제주</Option>
-            </Select>
-          </div>
-          <div className="relative flex w-full gap-2 w-100%">
-            <Input
-              type="text"
-              label="Search..."
-              defaultValue={inputText}
-              crossOrigin={undefined}
-              onChange={handlerInput}
-            />
-            <button
-              className="absolute right-0 m-2 text-xl cursor-pointer text-gray-3"
-              type="submit"
-              onClick={handleFormSubmit}
-            >
-              <AiOutlineSearch />
-            </button>
-          </div>
-        </form>
+        <SearchSelect searchList={searchList} setSearchList={setSearchList} />
+        <SearchForm
+          inputText={inputText}
+          setInputText={setInputText}
+          setSearchList={setSearchList}
+        />
       </Card>
     </>
   );
